Add FN0 for accessing "default" layer 0
Also rename "Fn layer" to "layer 1" and make corresponding changes to
constant names.  That drives changes to the save files so a new format
is necessary.  Finally, now that there is a FN0, the firmware logic has
to be updated to prevent it from trying to activate a "layer 0 active"
LED, which doesn't exist.